Fix for incorrect image deletion in Photo gallery module when preview was not yet created

+++ components/modules/Photo_gallery/Photo_gallery.php
@@ -247,7 +247,9 @@ class Photo_gallery {
 					'tag'	=> "Photo_gallery/images/$id"
 				]
 			);
-			$this->storage()->unlink($this->storage()->source_by_url($data['preview']));
+			if ($data['preview']) {
+				$this->storage()->unlink($this->storage()->source_by_url($data['preview']));
+			}
 			$Cache	= $this->cache;
 			unset(
 				$Cache->{"images/$id"},
